+++ examples/data-sources/clumio_aws_manual_connection_resources/data-source.tf
@@ -1,6 +1,6 @@
 data "clumio_aws_manual_connection_resources" "test_get_resources" {
-  account_native_id = "aws_account_id"
-  aws_region        = "aws_region"
+  account_native_id = "123456789012" # Replace with your actual AWS account ID.
+  aws_region        = "us-west-2" # Replace with your actual AWS region.
   asset_types_enabled = {
     ebs   = true
     rds   = true