+++ examples/resources/clumio_aws_connection/resource.tf
@@ -1,5 +1,5 @@
 resource "clumio_aws_connection" "example" {
-  account_native_id      = "aws_account_id"
-  aws_region             = "aws_region"
+  account_native_id      = "123456789012" # Replace with your actual AWS account ID.
+  aws_region             = "us-west-2" # Replace with your actual AWS region.
   description            = "description"
 }