+++ examples/resources/clumio_policy_rule/example_using_policy_resource.tf
@@ -27,5 +27,18 @@ resource "clumio_policy_rule" "example_1" {
   # Using the clumio_policy resource to get the id of the policy.
   policy_id      = clumio_policy.example_ebs.id
   before_rule_id = clumio_policy_rule.example_2.id
-  condition      = "{\"entity_type\":{\"$eq\":\"aws_ebs_volume\"}, \"aws_account_native_id\":{\"$in\":[\"aws_account_id_1\", \"aws_account_id_2\"]}, \"aws_tag\":{\"$eq\":{\"key\":\"aws_tag_key\", \"value\":\"aws_tag_value\"}}}"
+  condition = jsonencode({
+    "entity_type" : {
+      "$eq" : "aws_ebs_volume"
+    },
+    "aws_account_native_id" : {
+      "$in" : ["123456789012", "234567890123"]
+    },
+    "aws_tag" : {
+      "$eq" : {
+        "key" : "Key1",
+        "value" : "Value1"
+      }
+    }
+  })
 }